+++ src/main/java/com/synopsys/integration/detect/lifecycle/run/operation/blackduck/BinaryScanOperation.java
@@ -20,7 +20,7 @@ import com.synopsys.integration.detect.tool.binaryscanner.BlackDuckBinaryScanner
 import com.synopsys.integration.detect.workflow.codelocation.CodeLocationNameManager;
 import com.synopsys.integration.detect.workflow.event.EventSystem;
 import com.synopsys.integration.detect.workflow.file.DirectoryManager;
-import com.synopsys.integration.detectable.detectable.file.WildcardFileFinder;
+import com.synopsys.integration.common.util.finder.WildcardFileFinder;
 import com.synopsys.integration.exception.IntegrationException;
 import com.synopsys.integration.util.NameVersion;
