+++ src/main/java/com/synopsys/integration/detect/workflow/blackduck/ExclusionPatternCreator.java
@@ -17,7 +17,7 @@ import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.synopsys.integration.detectable.detectable.file.FileFinder;
+import com.synopsys.integration.common.util.finder.FileFinder;
 
 public class ExclusionPatternCreator {
     private final Logger logger = LoggerFactory.getLogger(ExclusionPatternCreator.class);
